requirements.txt

fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy